fix attach for scenes
Previously using `attach` on an `Object3D` would set the `parent` of the `Object3D`. Usually this has no bad effects, but if it is a `Scene` that will be rendered, it will cause an error in the render, because 3js uses the parent being null as a sentinel to stop ascending the tree during `updateMatrixWorld`.

I've moved the parent information for attachments into the `__r3f` object, like it is for non-`Object3D` things that are placed in the tree with `attach`, so this won't be a problem for `Scene`s or for anything else that shouldn't have a scene-graph parent.